Production-Ready RAG AI Agent/requirements.txt
streamlit
faiss-cpu
google-generativeai
sentence-transformers
PyPDF2
python-docx
langchain
Pillow
mimetypes
